proto: Clean up conversion warnings related to checksum macros
These should not change any behavior, they mostly clarify
the used types and silence warnings, since these casts are
deliberate.

Acked-by: MaxF <max@max-fillinger.net>
Gerrit URL: https://gerrit.openvpn.net/c/openvpn/+/1164
Message-Id: <20250926111726.153603-1-frank@lichtenheld.com>
URL: https://www.mail-archive.com/openvpn-devel@lists.sourceforge.net/msg33218.html

+++ src/openvpn/proto.h
@@ -214,7 +214,7 @@ struct ip_tcp_udp_hdr
  */
 #define ADJUST_CHECKSUM(acc, cksum)                \
     {                                              \
-        int _acc = acc;                            \
+        int32_t _acc = acc;                        \
         _acc += (cksum);                           \
         if (_acc < 0)                              \
         {                                          \
@@ -231,16 +231,16 @@ struct ip_tcp_udp_hdr
         }                                          \
     }
 
-#define ADD_CHECKSUM_32(acc, u32) \
-    {                             \
-        acc += (u32) & 0xffff;    \
-        acc += (u32) >> 16;       \
+#define ADD_CHECKSUM_32(acc, u32)         \
+    {                                     \
+        acc += (int32_t)((u32) & 0xffff); \
+        acc += (int32_t)((u32) >> 16);    \
     }
 
-#define SUB_CHECKSUM_32(acc, u32) \
-    {                             \
-        acc -= (u32) & 0xffff;    \
-        acc -= (u32) >> 16;       \
+#define SUB_CHECKSUM_32(acc, u32)         \
+    {                                     \
+        acc -= (int32_t)((u32) & 0xffff); \
+        acc -= (int32_t)((u32) >> 16);    \
     }
 
 /*
